requirements.txt
numpy
pymeterbus
